+++ app/controllers/collecting_events_controller.rb
@@ -146,7 +146,7 @@ class CollectingEventsController < ApplicationController
 
   # GET /collecting_events/download
   def download
-    send_data(Export::Download.generate_csv(CollectingEvent.where(project_id: sessions_current_project_id)),
+    send_data(Export::Csv.generate_csv(CollectingEvent.where(project_id: sessions_current_project_id)),
               type: 'text',
               filename: "collecting_events_#{DateTime.now}.tsv")
   end