+++ app/controllers/otus_controller.rb
@@ -236,7 +236,7 @@ class OtusController < ApplicationController
 
   # GET /otus/download
   def download
-    send_data Export::Download.generate_csv(Otu.where(project_id: sessions_current_project_id)),
+    send_data Export::Csv.generate_csv(Otu.where(project_id: sessions_current_project_id)),
       type: 'text',
       filename: "otus_#{DateTime.now}.tsv"
   end
@@ -250,6 +250,15 @@ class OtusController < ApplicationController
   # GET /otus/select_options?target=TaxonDetermination
   def select_options
     @otus = Otu.select_optimized(sessions_current_user_id, sessions_current_project_id, params.require(:target))
+  end
+
+  # POST /otus/batch_update.json?otus_query=<>&otu={taxon_name_id=123}}
+  def batch_update
+    c = Otu.batch_update({
+                           otu: otu_params.merge(by: sessions_current_user_id),
+                           otu_query: params[:otu_query]
+                         })
+    render json: c[:result], status: c[:status]
   end
 
   # GET /api/v1/otus.csv
@@ -266,7 +275,7 @@ class OtusController < ApplicationController
       }
       format.csv {
         @otus = q
-        send_data Export::Download.generate_csv(
+        send_data Export::Csv.generate_csv(
           @otus,
           exclude_columns: %w{updated_by_id created_by_id project_id},
         ), type: 'text', filename: "taxon_names_#{DateTime.now}.tsv"
@@ -282,6 +291,7 @@ class OtusController < ApplicationController
   def autocomplete
     @otus = ::Queries::Otu::Autocomplete.new(
       params.require(:term),
+      exact: 'true',
       project_id: sessions_current_project_id,
       with_taxon_name: params[:with_taxon_name],
       having_taxon_name_only: params[:having_taxon_name_only],
@@ -308,7 +318,7 @@ class OtusController < ApplicationController
   def api_dwc_inventory
     respond_to do |format|
       format.csv do
-        send_data Export::Download.generate_csv(
+        send_data Export::Csv.generate_csv(
           DwcOccurrence.scoped_by_otu(@otu),
           exclude_columns: ['id', 'created_by_id', 'updated_by_id', 'project_id', 'updated_at']),
           type: 'text',
@@ -406,7 +416,7 @@ class OtusController < ApplicationController
 # end
 
   def otu_params
-    params.require(:otu).permit(:name, :taxon_name_id)
+    params.require(:otu).permit(:name, :taxon_name_id, :exact)
   end
 
   def batch_params