+++ app/controllers/preparation_types_controller.rb
@@ -84,7 +84,7 @@ class PreparationTypesController < ApplicationController
 
   # GET /preparation_types/download
   def download
-    send_data(Export::Download.generate_csv(PreparationType.all),
+    send_data(Export::Csv.generate_csv(PreparationType.all),
               type: 'text',
               filename: "preparation_types_#{DateTime.now}.tsv")
   end