+++ app/javascript/vue/components/radials/co/radial.vue
@@ -31,6 +31,7 @@
                 :ids="ids"
                 :parameters="params"
                 :count="count"
+                @close="closeModal"
               />
             </div>
           </div>
@@ -61,6 +62,7 @@ import VBtn from '@/components/ui/VBtn/index.vue'
 import SliceTaxonDetermination from './components/SliceTaxonDetermination.vue'
 import SliceBiocurations from './components/SliceBiocurations/SliceBiocurations.vue'
 import SliceRepository from './components/SliceRepository.vue'
+import DwcSlice from './components/DwCSlice.vue'
 
 import { computed, ref } from 'vue'
 import { removeEmptyProperties } from '@/helpers/objects.js'
@@ -69,7 +71,8 @@ const EXCLUDE_PARAMETERS = ['per', 'page', 'extend']
 const SLICES = {
   'Add biocurations': SliceBiocurations,
   'Taxon determinations': SliceTaxonDetermination,
-  Repository: SliceRepository
+  Repository: SliceRepository,
+  'Regenerate DwC': DwcSlice
 }
 
 defineOptions({