+++ app/javascript/vue/tasks/contents/filter/App.vue
@@ -24,6 +24,7 @@
           :attributes="ATTRIBUTES"
           :list="list"
           @on-sort="list = $event"
+          @remove="({ index }) => list.splice(index, 1)"
         />
       </template>
     </FilterLayout>