+++ app/javascript/vue/tasks/loans/filter/App.vue
@@ -23,6 +23,7 @@
           :list="list"
           :attributes="ATTRIBUTES"
           @on-sort="list = $event"
+          @remove="({ index }) => list.splice(index, 1)"
         />
       </template>
     </FilterLayout>