+++ app/javascript/vue/tasks/nomenclature/filter/App.vue
@@ -41,6 +41,7 @@
           :attributes="ATTRIBUTES"
           :list="list"
           @on-sort="list = $event"
+          @remove="({ index }) => list.splice(index, 1)"
         />
       </template>
     </FilterLayout>