+++ app/javascript/vue/tasks/otu/browse/store/actions/loadInformation.js
@@ -5,6 +5,9 @@ import { TaxonName } from '@/routes/endpoints'
 export default async ({ dispatch, commit, state }, otus) => {
   const { currentOtu } = state
   const otuIds = otus.map((otu) => otu.id)
+  const taxonNameIds = [
+    ...new Set(otus.map((otu) => otu.taxon_name_id))
+  ].filter(Boolean)
 
   await dispatch(ActionNames.LoadPreferences)
 
@@ -22,11 +25,13 @@ export default async ({ dispatch, commit, state }, otus) => {
 
   dispatch(ActionNames.LoadDistribution, currentOtu.id)
 
-  TaxonName.all({
-    taxon_name_id: [...new Set(otus.map((otu) => otu.taxon_name_id))]
-  }).then((response) => {
-    commit(MutationNames.SetTaxonNames, response.body)
-  })
+  if (taxonNameIds.length) {
+    TaxonName.all({
+      taxon_name_id: taxonNameIds
+    }).then((response) => {
+      commit(MutationNames.SetTaxonNames, response.body)
+    })
+  }
 
   dispatch(ActionNames.LoadObservationDepictions, otus)
   dispatch(ActionNames.LoadDescendants, state.currentOtu)