+++ spec/models/biological_association_spec.rb
@@ -50,6 +50,104 @@ describe BiologicalAssociation, type: :model do
     )).to be_truthy
   end
 
+  specify '#batch_update' do
+    a = FactoryBot.create(:valid_biological_association)
+    b = FactoryBot.create(:valid_biological_association)
+    r = FactoryBot.create(:valid_biological_relationship)
+
+    BiologicalAssociation.batch_update(
+      biological_association: { biological_relationship_id: r.id},
+      biological_association_query: { }
+    )
+
+    expect(r.biological_associations.count).to eq(2)
+  end
+
+  specify '#batch_update preview' do
+    a = FactoryBot.create(:valid_biological_association)
+    b = FactoryBot.create(:valid_biological_association)
+    r = FactoryBot.create(:valid_biological_relationship)
+
+    BiologicalAssociation.batch_update(
+      preview: true,
+      biological_association: { biological_relationship_id: r.id},
+      biological_association_query: { }
+    )
+
+    expect(r.biological_associations.count).to eq(0)
+  end
+
+  specify '#batch_update errors' do
+    r = FactoryBot.create(:valid_biological_relationship)
+    
+    a = FactoryBot.create(:valid_biological_association, biological_relationship: r)
+    b = FactoryBot.create(:valid_biological_association, biological_relationship: r, biological_association_subject: a.biological_association_subject)
+
+    m = BiologicalAssociation.batch_update(
+      biological_association: { biological_relationship_id: r.id, biological_association_object_id: a.biological_association_object_id},
+      biological_association_query: { biological_association_id: [  b.id ] }
+    )
+
+    expect(m.errors).to eq( { "Validation failed: Biological association subject has already been taken" => 1 } )
+  end
+
+
+  specify '#batch_rotate' do
+    a = FactoryBot.create(:valid_biological_association)
+
+    c = a.subject
+    d = a.object
+
+    BiologicalAssociation.batch_rotate( {} )
+
+    a.reload
+    expect(a.subject).to eq(d)
+    expect(a.biological_association_object).to eq(c)
+  end
+
+  specify '#batch_rotate response' do
+    a = FactoryBot.create(:valid_biological_association)
+
+    c = a.subject
+    d = a.object
+
+    r = BiologicalAssociation.batch_rotate( {} )
+
+    expect(r.updated).to contain_exactly(a.id)
+    expect(r.not_updated).to eq([])
+    expect(r.errors).to eq({})
+  end
+
+    specify '#batch_rotate' do
+    a = FactoryBot.create(:valid_biological_association)
+
+    c = a.subject
+    d = a.object
+
+    BiologicalAssociation.batch_rotate( {} )
+
+    a.reload
+    expect(a.subject).to eq(d)
+    expect(a.biological_association_object).to eq(c)
+  end
+
+  specify '#batch_rotate preview' do
+    a = FactoryBot.create(:valid_biological_association)
+
+    c = a.subject
+    d = a.object
+
+    r = BiologicalAssociation.batch_rotate( { preview: true } )
+
+    expect(r.updated).to eq([a.id])
+    expect(r.not_updated).to eq([])
+    expect(r.preview).to be_truthy
+
+    a.reload
+    expect(a.subject).to eq(c)
+    expect(a.biological_association_object).to eq(d)
+  end
+
   context 'concerns' do
     it_behaves_like 'citations'
     it_behaves_like 'is_data'