+++ spec/models/dataset_record/darwin_core/taxon_spec.rb
@@ -1076,6 +1076,168 @@ describe 'DatasetRecord::DarwinCore::Taxon', type: :model do
 
   end
 
+  context 'when importing an original misspelling' do
+    before(:all) { import_checklist_tsv('original_misspelling.tsv', 6, 'misspelling combination') }
+
+    # Tetramorium rothneyi longi is a synonym of Tetramorium wroughtonii and has a different
+    # original combination (Rhoptromyrmex rothneyi longi). Because the valid rank is different and it is not the
+    # original combination, we need to create a subsequent combination for it.
+    # However, Tetramorium rothneyi is a subsequent combination (not a protonym) of Tetramorium wroughtonii rothneyi.
+    # This revealed a bug that subsequent combination generation for synonyms did not use the
+    # finest protonym for Combination ancestors.
+
+    after :all do
+      DatabaseCleaner.clean
+    end
+
+    let(:valid) { Protonym.find_by(cached: 'Polyrhachis (Myrma) olena') }
+    let(:misspelling) { Protonym.find_by({ name: 'olenus' }) }
+
+    it 'should create and import 5 records' do
+      verify_all_records_imported(6)
+    end
+
+    it 'should have a misspelling relationship ' do
+      relationship = TaxonNameRelationship.find_by({ subject_taxon_name_id: misspelling.id, object_taxon_name_id: valid.id })
+      expect(relationship.type_name).to eq('TaxonNameRelationship::Iczn::Invalidating::Usage::IncorrectOriginalSpelling')
+    end
+
+    # Polyrhachis [1x]
+    # Myrma [1x]
+    # Polyrhachis (Myrma) olena [3x]
+    # Polyrhachis (Myrma) olenus [3x]
+    it 'should have eight original combinations' do
+      expect(TaxonNameRelationship::OriginalCombination.all.length).to eq 8
+    end
+
+    it 'should be cached invalid' do
+      expect(misspelling.cached_is_valid).to be false
+    end
+
+    it 'the synonym should have cached valid taxon id' do
+      expect(misspelling.cached_valid_taxon_name_id).to eq valid.id
+    end
+  end
+
+  context 'when importing a synonym with a combination parent' do
+    before(:all) { import_checklist_tsv('synonym_combination_parent.tsv', 9) }
+
+    # Tetramorium rothneyi longi is a synonym of Tetramorium wroughtonii and has a different
+    # original combination (Rhoptromyrmex rothneyi longi). Because the valid rank is different and it is not the
+    # original combination, we need to create a subsequent combination for it.
+    # However, Tetramorium rothneyi is a subsequent combination (not a protonym) of Tetramorium wroughtonii rothneyi.
+    # This revealed a bug that subsequent combination generation for synonyms did not use the
+    # finest protonym for Combination ancestors.
+
+    after :all do
+      DatabaseCleaner.clean
+    end
+
+    let(:synonym_protonym) { Protonym.find_by(name: 'longi') }
+
+    it 'should create and import 11 records' do
+      verify_all_records_imported(11)
+    end
+
+    it 'the synonym should have species rank' do
+      expect(synonym_protonym.rank).to eq('species')
+    end
+
+    it 'the synonym should have the correct cached original combination' do
+      expect(synonym_protonym.cached_original_combination).to eq('Rhoptromyrmex rothneyi longi')
+    end
+
+    it 'the synonym subsequent combination should exist' do
+      expect(Combination.find_by(cached: 'Tetramorium rothneyi longi')).to be_truthy
+    end
+
+  end
+
+  context 'when importing a dataset without scientificNameAuthorship data' do
+    before(:all) { import_checklist_tsv('no_authorship.tsv', 2) }
+    after(:all) { DatabaseCleaner.clean }
+
+    it 'should import both names' do
+      verify_all_records_imported(2)
+    end
+
+    it 'neither protonym should have authorship data' do
+      expect(Protonym.find_by(name: 'Formicidae').author_string).to be_nil
+      expect(Protonym.find_by(name: 'Calyptites').author_string).to be_nil
+    end
+  end
+
+
+  context 'when importing a second time that builds off the first import' do
+    before(:all) do
+      DatabaseCleaner.start
+      import_dataset = ImportDataset::DarwinCore::Checklist.create!(
+        source: fixture_file_upload((Rails.root + 'spec/files/import_datasets/checklists/' + 'two_part_upload/two_part_upload_1.tsv'), 'text/plain'),
+        description: "part 1",
+        import_settings: {'use_existing_taxon_hierarchy' => false}
+      ).tap { |i| i.stage }
+
+      1.times { |_|
+        import_dataset.import(5000, 100)
+      }
+
+      import_dataset2 = ImportDataset::DarwinCore::Checklist.create!(
+        source: fixture_file_upload((Rails.root + 'spec/files/import_datasets/checklists/' + 'two_part_upload/two_part_upload_2.tsv'), 'text/plain'),
+        description: "part 2",
+        import_settings: {'use_existing_taxon_hierarchy' => true}
+      ).tap { |i| i.stage }
+
+      2.times { |_|
+        import_dataset2.import(5000, 100)
+      }
+    end
+
+    after :all do
+      DatabaseCleaner.clean
+    end
+
+    it 'should import 3 records' do
+      verify_all_records_imported(3)
+    end
+
+    it 'only two protonyms should be created' do
+      expect(TaxonName.where.not(name: 'Root').count).to eq 2
+    end
+
+  end
+
+  context 'when importing a taxon with invalid as a classification' do
+    before(:all) { import_checklist_tsv('invalid_as_classification.tsv', 3) }
+    after(:all) { DatabaseCleaner.clean }
+
+    let(:taxon) { Protonym.find_by(name: 'Cynipidae') }
+
+    it 'should import the 1 row' do
+      verify_all_records_imported(1)
+    end
+
+    it "should set the taxon's status to invalid" do
+      expect(taxon.taxon_name_classifications.map { |c| c.type_name }).to include 'TaxonNameClassification::Iczn::Available::Invalid'
+    end
+  end
+
+  context 'when importing a taxon with invalid as a relationship' do
+    before(:all) { import_checklist_tsv('invalid_as_relationship.tsv', 3) }
+    after(:all) { DatabaseCleaner.clean }
+
+    let(:valid_protonym) { Protonym.find_by(name: 'Biorhiza') }
+    let(:invalid_protonym) { Protonym.find_by(name: 'Biorhyza') }
+
+    it 'should import the 3 rows' do
+      verify_all_records_imported(3)
+    end
+
+    it 'the invalid taxon should have an invalid relationship with the valid taxon' do
+      expect_relationship(invalid_protonym, valid_protonym, 'TaxonNameRelationship::Iczn::Invalidating')
+    end
+
+  end
+
   # TODO test missing parent
   #
   # TODO test protonym is unavailable --- set classification on unsaved TaxonName